Massive update. Modified tracebuffer struct, overhauled test functions, added struct init functions, modified files to work with renamed k_defines.h, updated Makefile.
Tracebuffer struct had an array of trace structs, changed to array of pointers to
trace structs to match how structs are handled everywhere else in code. Initialize
for tracebuffer handles allocating space for all 16 traces in the tracebuffer struct.

The test functions have been overhauled to use asserts instead of visual inspection.
Rewrote the mains to test the same cases, but asserting on fails to make it easier
to tell right away if the structs still work.
NOTE: test_priority_queue still needs to be updated.

Added struct initialization functions to be used in testing functions as well as
during initialization.

Renamed defines and globals file to add the k_ prefix, so updated all source files
and Makefiles to reflect this change

Updated Makefile to compile everything as usual