Added atomic calls to the initial context of each process, more changes to come
The initial context of each process now has a call to atomicity to unmask
signals. For all processes the atomic count is set to 1. For normal processes
atomic(0) is called to unmask signals. For iprocesses, re-enabling atomicity
is left to the atomic(off) call in the signal handler. This works, but is
very limited. Need to add atomic count for each process to handle cases where
the atomic counts of two processes do not match.