Fixed wall clock by restructuring code. Most likely was an atomicity issue.
Changed the message_receive at the start and if statements to route based on
the message received to blocking until each message received right away. That seems
to have fixed the issue of blocking and disabling interrupts. Could be that this
method just avoids the bug so keep an eye out in the future.